+++ openapi_client/models/lizard_timeseries_rain.py
@@ -278,6 +278,7 @@ class LizardTimeseriesRain(object):
     def values(self):
         """Gets the values of this LizardTimeseriesRain.  # noqa: E501
 
+        Timeseries provided as a nested list. The inner list consists of exactly 2 values: timestamp, value  # noqa: E501
 
         :return: The values of this LizardTimeseriesRain.  # noqa: E501
         :rtype: list[list[float]]
@@ -288,6 +289,7 @@ class LizardTimeseriesRain(object):
     def values(self, values):
         """Sets the values of this LizardTimeseriesRain.
 
+        Timeseries provided as a nested list. The inner list consists of exactly 2 values: timestamp, value  # noqa: E501
 
         :param values: The values of this LizardTimeseriesRain.  # noqa: E501
         :type: list[list[float]]