+++ openapi_client/models/timeseries_rain.py
@@ -211,6 +211,7 @@ class TimeseriesRain(object):
     def values(self):
         """Gets the values of this TimeseriesRain.  # noqa: E501
 
+        Timeseries provided as a nested list. The inner list consists of exactly 2 values: timestamp, value  # noqa: E501
 
         :return: The values of this TimeseriesRain.  # noqa: E501
         :rtype: list[list[float]]
@@ -221,6 +222,7 @@ class TimeseriesRain(object):
     def values(self, values):
         """Sets the values of this TimeseriesRain.
 
+        Timeseries provided as a nested list. The inner list consists of exactly 2 values: timestamp, value  # noqa: E501
 
         :param values: The values of this TimeseriesRain.  # noqa: E501
         :type: list[list[float]]