+++ openapi_client/models/timeseries_rain_overview.py
@@ -216,6 +216,7 @@ class TimeseriesRainOverview(object):
     def values(self):
         """Gets the values of this TimeseriesRainOverview.  # noqa: E501
 
+        Timeseries provided as a nested list. The inner list consists of exactly 2 values: timestamp, value  # noqa: E501
 
         :return: The values of this TimeseriesRainOverview.  # noqa: E501
         :rtype: list[list[float]]
@@ -226,6 +227,7 @@ class TimeseriesRainOverview(object):
     def values(self, values):
         """Sets the values of this TimeseriesRainOverview.
 
+        Timeseries provided as a nested list. The inner list consists of exactly 2 values: timestamp, value  # noqa: E501
 
         :param values: The values of this TimeseriesRainOverview.  # noqa: E501
         :type: list[list[float]]